requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
